# Covenant/API/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "EasyPeasy API",
    "version": "v0.1"
  },
  "paths": {
...
    "/api/implanttemplates": {
      "get": {
        "tags": [
          "ImplantTemplateApi"
        ],
        "operationId": "GetImplantTemplates",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ImplantTemplate"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ImplantTemplateApi"
        ],
        "operationId": "CreateImplantTemplate",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ImplantTemplate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ImplantTemplate"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ImplantTemplateApi"
        ],
        "operationId": "EditImplantTemplate",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ImplantTemplate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ImplantTemplate"
            }
          }
        }
      }
    },
    "/api/implanttemplates/{id}": {
      "get": {
        "tags": [
          "ImplantTemplateApi"
        ],
        "operationId": "GetImplantTemplate",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ImplantTemplate"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ImplantTemplateApi"
        ],
        "operationId": "DeleteImplantTemplate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/implanttemplates/{name}": {
      "get": {
        "tags": [
          "ImplantTemplateApi"
        ],
        "operationId": "GetImplantTemplateByName",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ImplantTemplate"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ImplantLanguage": {
      "enum": [
        "CSharp"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ImplantLanguage",
        "modelAsString": false
      }
    },
    "CommunicationType": {
      "enum": [
        "HTTP",
        "SMB",
        "Bridge"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "CommunicationType",
        "modelAsString": false
      }
    },
    "ImplantDirection": {
      "enum": [
        "Push",
        "Pull"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ImplantDirection",
        "modelAsString": false
      }
    },
    "ListenerType": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "DotNetVersion": {
      "enum": [
        "Net35",
        "Net40",
        "NetCore31"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "DotNetVersion",
        "modelAsString": false
      }
    },
    "ImplantTemplate": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "language": {
          "$ref": "#/definitions/ImplantLanguage"
        },
        "commType": {
          "$ref": "#/definitions/CommunicationType"
        },
        "implantDirection": {
          "$ref": "#/definitions/ImplantDirection"
        },
        "compatibleListenerTypes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ListenerType"
          },
          "readOnly": true
        },
        "compatibleDotNetVersions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DotNetVersion"
          }
        },
        "stagerCode": {
          "type": "string"
        },
        "executorCode": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}